bot/requirements/base.txt
aiogram~=3.20.0
python-dotenv~=1.0.1
ujson~=5.10.0
cachetools~=5.5.2
tortoise-orm[asyncpg]~=0.24.0
tomli_w~=1.2.0
tomlkit~=0.13.2
groq==0.11.0
sentence-transformers==2.3.1
chromadb==0.4.22
PyPDF2==3.0.1
numpy<2.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.24.10
python-dotenv~=1.0.1
ujson~=5.10.0
cachetools~=5.5.2
tortoise-orm[asyncpg]~=0.24.0
tomli_w~=1.2.0
tomlkit~=0.13.2
groq==0.9.0
sentence-transformers==2.3.1
chromadb==0.4.22
PyPDF2==3.0.1
numpy<2.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.24.1
